# internal/docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Panahon API.",
        "title": "Panahon API",
        "contact": {
            "name": "Emilio Gozo",
            "email": "emiliogozo@proton.me"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/stations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "List stations",
                "parameters": [
                    {
                        "type": "string",
                        "name": "bbox",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "circle",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "limit",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedStations"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Create station",
                "parameters": [
                    {
                        "description": "Create station parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateStationReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Station"
                        }
                    }
                }
            }
        },
...
        "/stations/{station_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Get station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station ID",
                        "name": "station_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Station"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Update station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station ID",
                        "name": "station_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update station parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateStationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Station"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Delete station",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Station ID",
                        "name": "station_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateStationReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_installed": {
                    "type": "string"
                },
                "elevation": {
                    "type": "number"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "mobile_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "station_type": {
                    "type": "string"
                },
                "station_type2": {
                    "type": "string"
                },
                "station_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "PaginatedStations": {
            "type": "object"
        },
...
        "Station": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_installed": {
                    "type": "string"
                },
                "elevation": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "mobile_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "station_type": {
                    "type": "string"
                },
                "station_type2": {
                    "type": "string"
                },
                "station_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "UpdateStationReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_installed": {
                    "type": "string"
                },
                "elevation": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "mobile_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "station_type": {
                    "type": "string"
                },
                "station_type2": {
                    "type": "string"
                },
                "station_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}